# services/payment-core/api/openapi.yaml
openapi: 3.0.3
info:
  title: payment-core API
  version: 1.0.0
  description: Core payment engine for Paylane handling payment state machine, double-entry ledger, provider adapters, webhooks, and outbox.
servers:
  - url: http://payment-core:4011
paths:
...
  /payments:
    post:
      summary: Initiate payment session (idempotent)
      security:
        - JWEAuth: []
      responses:
        '201':
          description: Payment session created or existing returned
        '400':
          description: Invalid request parameters
        '401':
          description: Unauthorized JWE
  /payments/{id}:
    get:
      summary: Retrieve payment by ULID
      security:
        - JWEAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment details
        '404':
          description: Payment not found
...
components:
  securitySchemes:
    JWEAuth:
      type: http
      scheme: bearer
      bearerFormat: JWE